Renamed packages and Added Svg interface issue_2
