+++ src/main/View/Drawing.java
@@ -1,3 +1,4 @@
+package View;
 public class Drawing {
 
 	public static void main(String[] args) {
